Refactor project into smaller modules and add tests
Split the code for converting imports and exports into smaller modules
and add some basic unit tests using mocha.
