+++ packages/convert-to-es-modules/package.json
@@ -14,5 +14,14 @@
     "glob": "^7.1.6",
     "recast": "^0.18.5",
     "resolve": "^1.14.2"
+  },
+  "scripts": {
+    "format": "prettier --write **/*.js",
+    "test": "mocha"
+  },
+  "devDependencies": {
+    "chai": "^4.2.0",
+    "mocha": "^7.0.0",
+    "prettier": "1.19.1"
   }
 }
